+++ client/src/test/suite/extension.test.ts
@@ -1010,239 +1010,4 @@ suite('Push Button Test Suite', function (this: Suite) {
     pushStub.reset()
     showInfoMsgStub.reset()
   })
-  test('validateContent only runs when it should', async () => {
-    const showInformationMessageStub = sinon.stub(vscode.window, 'showInformationMessage')
-    const openAndValidateStub = sinon.stub(pushContent, 'openAndValidate')
-
-    showInformationMessageStub.resolves(undefined)
-    await pushContent.validateContent()
-    assert(openAndValidateStub.notCalled)
-    openAndValidateStub.reset()
-
-    // 'as any' is required here because of showQuickPick overloading
-    showInformationMessageStub.resolves(pushContent.DocumentsToOpen.all as any)
-    await pushContent.validateContent()
-    assert(openAndValidateStub.calledWith(pushContent.DocumentsToOpen.all))
-  })
-  test('getDocumentsToOpen returns all files', async () => {
-    const findFilesStub = sinon.stub(vscode.workspace, 'findFiles')
-    const filesToReturn = [vscode.Uri.file('/a'), vscode.Uri.file('/b'), vscode.Uri.file('/c')]
-    const openDocuments = [vscode.Uri.file('/b').toString()]
-    findFilesStub.resolves(filesToReturn)
-    let toOpen = await pushContent.getDocumentsToOpen(pushContent.DocumentsToOpen.all, new Set())
-    assert(findFilesStub.calledOnce)
-    filesToReturn.forEach(uri => {
-      assert(toOpen.has(uri.toString()))
-    })
-
-    // We do not need to open documents that are already open
-    toOpen = await pushContent.getDocumentsToOpen(
-      pushContent.DocumentsToOpen.all,
-      new Set(openDocuments)
-    )
-    assert(!toOpen.has(openDocuments[0]))
-  })
-  test('getDocumentsToOpen returns changed files', async () => {
-    const filesToReturn = [
-      { uri: vscode.Uri.file('/a') },
-      { uri: vscode.Uri.file('/b') },
-      { uri: vscode.Uri.file('/c') }
-    ]
-    const openDocuments = [vscode.Uri.file('/b').toString()]
-    const diffWithHEADStub = sinon.stub()
-    const stubRepo = {
-      diffWithHEAD: diffWithHEADStub
-    } as any as Repository
-    sinon.stub(pushContent, 'getRepo').returns(stubRepo)
-
-    diffWithHEADStub.resolves(filesToReturn)
-    let toOpen = await pushContent.getDocumentsToOpen(pushContent.DocumentsToOpen.modified, new Set())
-    assert(diffWithHEADStub.calledOnce)
-    filesToReturn.forEach(o => {
-      assert(toOpen.has(o.uri.toString()))
-    })
-
-    // We do not need to open documents that are already open
-    toOpen = await pushContent.getDocumentsToOpen(
-      pushContent.DocumentsToOpen.modified,
-      new Set(openDocuments)
-    )
-    assert(!toOpen.has(openDocuments[0]))
-  })
-  test('getOpenDocuments returns expected values', async () => {
-    const activeTextEditorStub = sinon.stub(vscode.window, 'activeTextEditor')
-    const executeCommandStub = sinon.stub(vscode.commands, 'executeCommand')
-
-    activeTextEditorStub.get(() => undefined)
-    let openDocuments = await pushContent.getOpenDocuments()
-    assert(executeCommandStub.notCalled)
-    assert.strictEqual(openDocuments.size, 0)
-    executeCommandStub.reset()
-    activeTextEditorStub.reset()
-
-    // The expected behavior is for workbench.action.nextEditor to loop around to the first editor.
-    // After it loops around, the document that getOpenDocuments started on will be added to the
-    // set and the function will return.
-    activeTextEditorStub.get(() => {
-      // NOTE: executeCommandStub could be called with something other than nextEditor
-      // Could this cause unexpected behavior? Stay tuned to find out!
-      switch (executeCommandStub.callCount) {
-        case 0:
-        case 3:
-          return { document: { uri: vscode.Uri.file('/a') } }
-        case 1:
-          return { document: { uri: vscode.Uri.file('/b') } }
-        case 2:
-          return { document: { uri: vscode.Uri.file('/c') } }
-        default:
-          throw new Error('Something went wrong when looking for documents')
-      }
-    })
-    executeCommandStub.resolves()
-    openDocuments = await pushContent.getOpenDocuments()
-    assert.strictEqual(openDocuments.size, 3)
-    assert.strictEqual(executeCommandStub.callCount, 3)
-  })
-  test('getOpenDocuments or openAndValidate canceled', async () => {
-    const activeTextEditorStub = sinon.stub(vscode.window, 'activeTextEditor')
-    const executeCommandStub = sinon.stub(vscode.commands, 'executeCommand')
-    const withProgressStub = sinon.stub(vscode.window, 'withProgress')
-    let error: Error | undefined
-    activeTextEditorStub.get(() => ({ document: { uri: vscode.Uri.file('/a') } }))
-    // I could stub vscode.CancellationSource, but this seems less error prone
-    withProgressStub.callsFake((
-      options: vscode.ProgressOptions,
-      task: (
-        progress: vscode.Progress<{ message?: string, increment?: number }>,
-        token: vscode.CancellationToken
-      ) => Thenable<unknown>
-    ): Thenable<unknown> => {
-      return new Promise((resolve, reject) => {
-        try {
-          resolve(task(
-            { report: (_: { message?: string, increment?: number }) => {} },
-            { isCancellationRequested: true, onCancellationRequested: sinon.stub() }
-          ))
-        } catch (e) {
-          reject(e)
-        }
-      })
-    })
-    try {
-      await pushContent.getOpenDocuments()
-    } catch (e) {
-      error = e as Error
-    }
-    assert(withProgressStub.calledOnce)
-    assert(executeCommandStub.notCalled)
-    assert.notStrictEqual(error, undefined)
-    assert.strictEqual(error?.message, 'Canceled')
-    withProgressStub.resetHistory()
-    executeCommandStub.reset()
-    error = undefined
-
-    const getOpenDocumentsStub = sinon.stub(pushContent, 'getOpenDocuments')
-    const getDocumentsToOpenStub = sinon.stub(pushContent, 'getDocumentsToOpen')
-    getOpenDocumentsStub.resolves(new Set())
-    getDocumentsToOpenStub.resolves(new Set(['not', 'used', 'here']))
-    try {
-      await pushContent.openAndValidate(pushContent.DocumentsToOpen.modified)
-    } catch (e) {
-      error = e as Error
-    }
-    assert(withProgressStub.calledOnce)
-    assert(executeCommandStub.notCalled)
-    assert.notStrictEqual(error, undefined)
-    assert.strictEqual(error?.message, 'Canceled')
-  })
-  test('openAndValidate integration test', async () => {
-    const dateNowStub = sinon.stub(Date, 'now')
-    const withProgressStub = sinon.stub(vscode.window, 'withProgress')
-    const getOpenDocumentsStub = sinon.stub(pushContent, 'getOpenDocuments')
-    const getDocumentsToOpenStub = sinon.stub(pushContent, 'getDocumentsToOpen')
-    const showTextDocumentStub = sinon.stub(vscode.window, 'showTextDocument')
-      .callsFake((uri: vscode.Uri, options?: vscode.TextDocumentShowOptions): Thenable<vscode.TextEditor> => {
-        return new Promise((resolve, reject) => resolve(
-          { document: { uri: uri } as any as vscode.TextDocument } as any as vscode.TextEditor
-        ))
-      })
-    const executeCommandStub = sinon.stub(vscode.commands, 'executeCommand').resolves()
-    const filesToReturn = [vscode.Uri.file('/a'), vscode.Uri.file('/b'), vscode.Uri.file('/c')]
-    let dateNowCallCount = 0
-    let progressReportCount = 0
-    sinon.stub(pushContent, 'sleep').resolves()
-    // Cover situations that take more than 10 seconds
-    dateNowStub.callsFake(() => dateNowCallCount++ * 10000)
-    getOpenDocumentsStub.resolves(new Set())
-    getDocumentsToOpenStub.resolves(new Set(filesToReturn.map(uri => uri.toString())))
-    withProgressStub.callsFake((
-      options: vscode.ProgressOptions,
-      task: (
-        progress: vscode.Progress<{ message?: string, increment?: number }>,
-        token: vscode.CancellationToken
-      ) => Thenable<unknown>
-    ): Thenable<unknown> => {
-      return new Promise((resolve, reject) => {
-        try {
-          resolve(task(
-            {
-              report: (value: { message?: string, increment?: number }) => {
-                progressReportCount++
-                assert(
-                  value.message !== undefined &&
-                  value.message.length > 0 && (
-                    // make sure the time estimate is only added after the first progress report
-                    progressReportCount > 1
-                      ? value.message.includes('remaining')
-                      : !value.message.includes('remaining')
-                  )
-                )
-              }
-            },
-            { isCancellationRequested: false, onCancellationRequested: sinon.stub() }
-          ))
-        } catch (e) {
-          reject(e)
-        }
-      })
-    })
-
-    let errors = await pushContent.openAndValidate(pushContent.DocumentsToOpen.all)
-    assert.strictEqual([...errors.values()].flat().length, 0)
-    filesToReturn.forEach(uri => {
-      assert(showTextDocumentStub.calledWith(uri))
-    })
-    assert(getOpenDocumentsStub.called)
-    assert(executeCommandStub.calledWith('workbench.action.closeActiveEditor'))
-    assert.strictEqual(executeCommandStub.callCount, 3) // Close three documents with no errors
-    assert.strictEqual(dateNowStub.callCount, 11)
-    assert.strictEqual(withProgressStub.callCount, 1)
-    assert.strictEqual(progressReportCount, 4) // 1 extra call to get the progress bar spinning
-    getOpenDocumentsStub.resetHistory()
-    executeCommandStub.resetHistory()
-    withProgressStub.resetHistory()
-    showTextDocumentStub.resetHistory()
-    progressReportCount = 0
-    dateNowStub.reset()
-
-    // Test for cases where errors appear (when documents should not be closed)
-    const file1Diag1 = { severity: vscode.DiagnosticSeverity.Error, source: 'source1' } as any as vscode.Diagnostic
-    const testDiagnostics: Array<[vscode.Uri, vscode.Diagnostic[]]> = [
-      [filesToReturn[0], [file1Diag1]]
-    ]
-    // Cover situations that take a very small amount of time
-    dateNowStub.callThrough()
-    sinon.stub(vscode.languages, 'getDiagnostics').returns(testDiagnostics)
-    errors = await pushContent.openAndValidate(pushContent.DocumentsToOpen.all)
-    assert.strictEqual([...errors.values()].flat().length, 1)
-    filesToReturn.forEach(uri => {
-      assert(showTextDocumentStub.calledWith(uri))
-    })
-    assert(getOpenDocumentsStub.called)
-    assert(executeCommandStub.calledWith('workbench.action.closeActiveEditor'))
-    assert.strictEqual(executeCommandStub.callCount, 2) // Close two documents with no errors
-    assert.strictEqual(dateNowStub.callCount, 5)
-    assert.strictEqual(withProgressStub.callCount, 1)
-    assert.strictEqual(progressReportCount, 1) // Just the 1 to get the progress bar spinning
-  })
 })
